Rearrange directories: move ContentArea in its own directory
