+++ frontend/src/ContentArea/ContentArea.js
